[tests] Exercise the ZIP reader with malformed archives
The reader is pure byte parsing over untrusted input, so it can be driven
entirely from memory: build an archive byte by byte, hand it to open_memory and
say what should come back. Covers a well-formed stored entry, a directory
promising more entries than it holds and one promising the largest count the
record can express, a directory whose names do not fit the arena its entry count
leaves room for, an extra field record that overruns its field, a stored entry
whose two sizes disagree, and, over the entry streams, reading a stored and a
deflated entry back byte for byte, rewinding each after a partial read, and an
entry whose deflate stream ends before the size it declared.

Adds the reader and both entry streams to the test target, which needs zlib
linked in. Compiling them with the suite's -Wall -Wextra surfaced a signed
comparison in the entry storage assertion; make the bound an int.

+++ src/bstone/src/bstone_zip_archive_file.cpp
@@ -596,7 +596,7 @@ bool ZipArchiveFile::initialize_entry_name(const CentralFileHeader& header, Arch
 void ZipArchiveFile::reserve_entries(int capacity)
 {
 	BSTONE_ASSERT(capacity >= 0);
-	BSTONE_ASSERT(capacity <= INT_MAX / sizeof(ArchiveFileEntry));
+	BSTONE_ASSERT(capacity <= static_cast<int>(INT_MAX / sizeof(ArchiveFileEntry)));
 	if (entries_capacity_ >= capacity)
 		return;
 	const int storage_size = capacity * sizeof(ArchiveFileEntry);
